move gcp formatter to different file
